Refactor models to module files
