+++ core/models/__init__.py
@@ -1,9 +1,22 @@
-from .models import (
-    VLAN,
+from modules.inventory.models import (
     Device,
-    SSHCredential,
-    SNMPCommunity,
     DeviceType,
+    DeviceEditLog,
+    DeviceDamage,
+    Tag,
+    Location,
+)
+from modules.network.models import (
+    VLAN,
+    SNMPCommunity,
+    SSHCredential,
+    Interface,
+    InterfaceChangeLog,
+    PortStatusHistory,
+    PortConfigTemplate,
+    ConnectedSite,
+)
+from .models import (
     Site,
     SiteMembership,
     ConfigBackup,
@@ -11,15 +24,9 @@ from .models import (
     UserSSHCredential,
     SystemTunable,
     AuditLog,
-    PortConfigTemplate,
-    Tag,
-    DeviceEditLog,
     BannedIP,
     LoginEvent,
     EmailLog,
-    PortStatusHistory,
-    Interface,
-    InterfaceChangeLog,
     DashboardWidget,
     SiteDashboardWidget,
     SiteKey,
@@ -42,8 +49,10 @@ __all__ = [
     "SystemTunable",
     "AuditLog",
     "PortConfigTemplate",
+    "DeviceDamage",
     "Tag",
     "DeviceEditLog",
+    "Location",
     "BannedIP",
     "LoginEvent",
     "EmailLog",
@@ -52,6 +61,7 @@ __all__ = [
     "InterfaceChangeLog",
     "DashboardWidget",
     "SiteDashboardWidget",
+    "ConnectedSite",
     "SiteKey",
     "CustomColumn",
     "ColumnPreference",
