+++ modules/inventory/__init__.py
@@ -1,5 +1,4 @@
 """Inventory module exports."""
 
-from .routes import router as inventory_router
-
-__all__ = ["inventory_router"]
+# Router is defined in modules.inventory.routes
+__all__ = []
