+++ server/routes/internal/snmp_traps.py
@@ -5,7 +5,8 @@ from core.utils.db_session import get_db
 from sqlalchemy.orm import Session
 from core.utils.auth import require_role
 from core.utils.templates import templates
-from core.models.models import SNMPTrapLog, Device
+from modules.inventory.models import Device
+from core.models.models import SNMPTrapLog
 
 router = APIRouter()
