+++ server/routes/ui/add_device.py
@@ -6,7 +6,7 @@ import csv
 from core.utils.templates import templates
 from core.utils.auth import require_role
 from core.utils.db_session import get_db, reset_pk_sequence
-from core.models.models import Device, DeviceType, Location
+from modules.inventory.models import Device, DeviceType, Location
 from server.routes.ui.task_views import _open_sheet
 from server.routes.ui.devices import _format_ip
