+++ tests/test_damage_upload.py
@@ -29,7 +29,7 @@ def get_test_app():
 def test_upload_damage_photo_sanitizes_filename(tmp_path, monkeypatch):
     app = get_test_app()
     from server.routes import devices as devices_module
-    from core.models.models import DeviceDamage, Device
+    from modules.inventory.models import Device, DeviceDamage
 
     # Prepare dummy device and DB session
     dummy_device = types.SimpleNamespace(id=1, site_id=1)
